Make groovydoc attach artifact (used in releases) less intrusive
Previously, any project that used build-helper-maven-plugin for any
other objective triggered this execution in the package phase during release.
This would then trigger a cascading failure on gpg signing later on:
  gpg: impossible d'ouvrir « /tmp/powsybl-parent/powsybl-parent/powsybl-parent-ws/target/powsybl-parent-ws-5-groovydoc.jar » : Aucun fichier ou dossier de ce type
  gpg: signing failed: erreur d'ouverture de fichier

Projects not using groovy needed to turn it off. Now projects using groovy need to turn it on.

This has the disadvantage of copy pasting the phase everywhere but it shouldn't be a problem (the phase will never change)
This has the added advantage that the plugin execution is more explicit now because projects
have to add the "attach-groovydoc-artifactId" id, instead of just enabling all build-helper executions
diff:
---
  <plugin>
      <groupId>org.codehaus.mojo</groupId>
      <artifactId>build-helper-maven-plugin</artifactId>
+     <executions>
+       <execution>
+         <id>attach-groovydoc-artifact</id>
+         <phase>package</phase>
+       </execution>
+     </executions>
  </plugin>
---

Signed-off-by: HARPER Jon <jon.harper87@gmail.com>